+++ src/mlp.py
@@ -10,7 +10,7 @@ import slope_interval
 from auto_LiRPA import BoundedModule, BoundedTensor
 from auto_LiRPA.perturbations import PerturbationLpNorm
 
-torch.set_default_tensor_type(torch.cuda.FloatTensor)
+torch.set_default_tensor_type(torch.cuda.DoubleTensor)
 device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
 
 
@@ -173,8 +173,8 @@ def func_as_torch(params):
         # print(name)
         # self.op_list.append((name, args))
         if name == 'dense':
-            A = torch.tensor(args['A']).T#.to(device)
-            b = torch.tensor(args['b'])#.to(device)
+            A = torch.tensor(args['A'], dtype=torch.double).T#.to(device)
+            b = torch.tensor(args['b'], dtype=torch.double)#.to(device)
             linear = torch.nn.Linear(A.shape[1], A.shape[0])
             linear.weight = torch.nn.Parameter(A)
             linear.bias = torch.nn.Parameter(b)
@@ -182,9 +182,7 @@ def func_as_torch(params):
         elif name == 'relu':
             op_list.append(torch.nn.ReLU())
     model = torch.nn.Sequential(*op_list)
-    # model.to(device)
-    # print("Torch Model: ")
-    # print(model)
+
     return model
